# services/product-service/docs/swagger.yaml
definitions:
  models.ProductForm:
    properties:
      detail:
        type: string
      name:
        type: string
      picture:
        type: string
      price:
        type: number
    type: object
info:
  contact: {}
paths:
  /prod/Add/:
    post:
      consumes:
      - application/json
      description: Post new product
      parameters:
      - description: New Product Form
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProductForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Post new product
      tags:
      - products
  /prod/DeleteById/{id}:
    delete:
      consumes:
      - application/json
      description: Delete product by id
      parameters:
      - description: Product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Delete product by id
      tags:
      - products
  /prod/EditById/{id}:
    put:
      consumes:
      - application/json
      description: Edit existing product by id
      parameters:
      - description: Product Id
        in: path
        name: id
        required: true
        type: string
      - description: Edit Product Form
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProductForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Edit existing product by id
      tags:
      - products
  /prod/GetById/{id}:
    get:
      consumes:
      - application/json
      description: Get product by id
      parameters:
      - description: Product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get product by id
      tags:
      - products
  /prod/GetByOwner/{id}:
    get:
      consumes:
      - application/json
      description: Get product by owner id
      parameters:
      - description: Owner Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get product by owner id
      tags:
      - products
swagger: "2.0"